Update flags.cmake
The error is caused by the -m64 flag, which, as far as I know, is mainly used in Windows environments. Not in Linux AARCH64 systems such as the RK3399, Raspberry Pi 64-bit OS or Jetson Nano, TX or Xavier. They all don't recognize the -m64 flag and therefore fall into the same error. Best the skip the setting of the flag when using an AARCH64 OS. Can be done with modifying line 151 in   ##Paddle-Lite/cmake/flags.cmake
```
if(NOT LITE_WITH_CUDA AND NOT(CMAKE_SYSTEM_PROCESSOR MATCHES "^(aarch64.*|AARCH64.*)"))
    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -m64")
endif()
```