RTOS support: support paddle lite run on zephyr rtos
We integrated Paddle Lite as a third-party external module into the Zephyr RTOS code repository.
Subsequently, we conducted tests on the Firefly ROC-RK3568-PC development board.
The predictive outcomes aligned with those of ARM Linux. Additionally, in some scenarios,
the speed of Paddle Lite on Zephyr RTOS surpassed that of Linux.

Signed-off-by: Charlie Xiong <1981639884@qq.com>